Add Dockerfile and setup script for configuring SFTP users
This commit adds a Dockerfile and a setup script for configuring SFTP users in the container. The Dockerfile is based on the atmoz/sftp image and copies the users.conf file to /etc/sftp/users.conf. It also copies the setup-users.sh script to /usr/local/bin/setup-users.sh and makes it executable. The script reads the users.conf file and creates the specified users with their passwords, UIDs, GIDs, and home directories. Finally, the script starts the SFTP process of the original container.

Add package.json with start and dev scripts

This commit adds a package.json file with the necessary scripts for starting the Railapp SFTP server. The "start" script runs the index.js file using the "node" command, while the "dev" script uses "nodemon" for automatic restarts during development.

Add index.js file with initial console log

This commit adds an index.js file with a single console log statement. This serves as the initial code for the Railapp SFTP server.

Add users.conf file with sample user configurations

This commit adds a users.conf file with sample user configurations. Each line in the file specifies a username, password, UID, GID, and home directory for an SFTP user.